Arquivo conftest.py incluído dentro de test_spam para disponibilizar as fixtures entre diferentes módulos.
