+++ core/tests/test_spam/test_usuarios.py
@@ -1,26 +1,4 @@
-import pytest
-
-from core.spam.db import Conexao
 from core.spam.modelos import Usuario
-
-
-@pytest.fixture(scope='module')
-def conexao():
-    # SetUp
-    conexao_obj = Conexao()
-    yield conexao_obj
-    # Tear Down
-    conexao_obj.fechar()
-
-
-@pytest.fixture
-def sessao(conexao):
-    # SetUp
-    sessao_obj = conexao.gerar_sessao()
-    yield sessao_obj
-    # Tear Down
-    sessao_obj.roll_back()
-    sessao_obj.fechar()
 
 
 def test_salvar_usuario(sessao):
